chore: update dependencies, restrict LinkML to >=1.9.2, <1.9.5
The Poetry lock file was stale, but running `poetry update` seemed to be
causing tests to fail. This was because the previous, looser upper
bound led to the fetch of the latest LinkML v1.9.6, which introduces a
regression, at least in our data, for e.g. in
`resources/schema/mapping_package_v3/models/context.jsonld`:

```
       "document_version_range": {
-         "@type": "mfy:VersionRange",
+         "@type": "@id",
          "@id": "document_version_range"
       },
```

This seemed totally nonsensical, and upon further investigation, this
appears to manifest from 1.9.5 onwards. While a search on their issue
tracker is pending (a brief glance does not yield any clue or it looks
like no-one has reportd it), we can restrict our dependency version
range for now.